requirements.txt
anthropic
openai
httpx
beautifulsoup4
PyPDF2
readability-lxml
textstat
lxml
